clean up router package
